user authorization

+++ backend/routes/posts.js
@@ -35,6 +35,7 @@ router.post(
       title: req.body.title,
       content: req.body.content,
       imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then((createdPost) => {
       res.status(201).json({
@@ -82,8 +83,14 @@ router.get("/:id", (req, res) => {
 });
 
 router.delete("/:id", checkAuth, (req, res) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    res.status(200).json({ message: "post deleted!" });
+  Post.deleteOne({ _id: req.params.id , creator: req.userData.userId}).then((result) => {
+    console.log(result);
+    if(result.deletedCount > 0){
+      res.status(200).json({ message: "post deleted!" });
+    }
+    else{
+      res.status(401).json({message : "unauthorized error"})
+    }
   });
 });
 
@@ -102,10 +109,15 @@ router.put(
       title: req.body.title,
       content: req.body.content,
       imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      console.log(result);
-      res.status(200).json(result);
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId}, post).then((result) => {
+      if(result.modifiedCount > 0){
+        res.status(200).json(result);
+      }
+      else{
+        res.status(401).json({message : "unauthorized error"})
+      }
     });
   }
 );
